+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/signaling/message/Catalog.kt
@@ -67,11 +67,11 @@ data class CandidateMessage(
 
 data class NotificationMessage(
         @SerializedName("event_type")                     val eventType:                     String,
-        @SerializedName("role")                           val role:                          String,
-        @SerializedName("minutes")                        val connectionTime:                Long,
-        @SerializedName("channel_connections")            val numberOfConnections:           Int,
-        @SerializedName("channel_upstream_connections")   val numberOfUpstreamConnections:   Int,
-        @SerializedName("channel_downstream_connections") val numberOfDownstreamConnections: Int,
+        @SerializedName("role")                           val role:                          String?,
+        @SerializedName("minutes")                        val connectionTime:                Long?,
+        @SerializedName("channel_connections")            val numberOfConnections:           Int?,
+        @SerializedName("channel_upstream_connections")   val numberOfUpstreamConnections:   Int?,
+        @SerializedName("channel_downstream_connections") val numberOfDownstreamConnections: Int?,
         @SerializedName("audio")                          val audio:                         Boolean?,
         @SerializedName("video")                          val video:                         Boolean?,
         @SerializedName("spotlight_id")                   val spotlightId:                   String?